requirements.txt
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
requests==2.32.3
requests-file==2.1.0
tldextract==5.3.0
ua-generator==2.0.6
urllib3==2.4.0
Werkzeug==3.1.3

